+++ src/main/java/net/jptrzy/inventory/backpack/client/renderer/EnderBackpackArmorRenderer.java
@@ -4,7 +4,6 @@ import net.fabricmc.fabric.api.client.rendering.v1.ArmorRenderer;
 import net.jptrzy.inventory.backpack.Main;
 import net.jptrzy.inventory.backpack.client.model.BackpackModel;
 import net.jptrzy.inventory.backpack.item.BackpackItem;
-import net.minecraft.client.model.ModelPart;
 import net.minecraft.client.render.OverlayTexture;
 import net.minecraft.client.render.RenderLayer;
 import net.minecraft.client.render.VertexConsumer;
@@ -17,9 +16,9 @@ import net.minecraft.entity.LivingEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.Identifier;
 
-public class BackpackArmorRenderer implements ArmorRenderer {
+public class EnderBackpackArmorRenderer implements ArmorRenderer {
 
-    public static final Identifier BACKPACK_MODEL_TEXTURE = Main.id("textures/armor/backpack.png");
+    public static final Identifier BACKPACK_MODEL_TEXTURE = Main.id("textures/armor/ender_backpack.png");
     private BackpackModel model;
 
     @Override
@@ -27,15 +26,10 @@ public class BackpackArmorRenderer implements ArmorRenderer {
         BackpackModel model = getModel();
         contextModel.setAttributes(model);
 
-        int i = ((BackpackItem) stack.getItem()).getColor(stack);
-        float r = (float) (i >> 16 & 255) / 255.0F;
-        float g = (float) (i >> 8 & 255) / 255.0F;
-        float b = (float) (i & 255) / 255.0F;
-
         model.setAngles(entity, 0.0F, 0.0F, 0.0F, 0.0F, 0.0F);
 
         VertexConsumer vertexConsumer = ItemRenderer.getArmorGlintConsumer(vertexConsumers, RenderLayer.getArmorCutoutNoCull(BACKPACK_MODEL_TEXTURE), false, stack.hasGlint());
-        model.render(matrices, vertexConsumer, light, OverlayTexture.DEFAULT_UV, r, g, b, 1);
+        model.render(matrices, vertexConsumer, light, OverlayTexture.DEFAULT_UV, 1, 1, 1, 1);
     }
 
     public BackpackModel getModel() {