+++ src/main/java/net/jptrzy/inventory/backpack/item/BackpackArmorMaterial.java
@@ -1,0 +1,54 @@
+package net.jptrzy.inventory.backpack.item;
+
+import net.minecraft.entity.EquipmentSlot;
+import net.minecraft.item.ArmorMaterial;
+import net.minecraft.item.ArmorMaterials;
+import net.minecraft.recipe.Ingredient;
+import net.minecraft.sound.SoundEvent;
+import net.minecraft.sound.SoundEvents;
+
+public class BackpackArmorMaterial implements ArmorMaterial {
+    private static final int[] BASE_DURABILITY = new int[] {13, 15, 16, 11};
+    private static final int[] PROTECTION_VALUES = new int[] {0, 0, 0, 0};
+
+    @Override
+    public int getDurability(EquipmentSlot slot) {
+        return BASE_DURABILITY[slot.getEntitySlotId()];
+    }
+
+    @Override
+    public int getProtectionAmount(EquipmentSlot slot) {
+        return PROTECTION_VALUES[slot.getEntitySlotId()];
+    }
+
+    @Override
+    public int getEnchantability() {
+        return 0;
+    }
+
+    @Override
+    public SoundEvent getEquipSound() {
+        return ArmorMaterials.LEATHER.getEquipSound();
+    }
+
+    @Override
+    public Ingredient getRepairIngredient() {
+        return ArmorMaterials.LEATHER.getRepairIngredient();
+    }
+
+    @Override
+    public String getName() {
+        // Must be all lowercase
+        return "backpack";
+    }
+
+    @Override
+    public float getToughness() {
+        return 0;
+    }
+
+    @Override
+    public float getKnockbackResistance() {
+        return 0;
+    }
+}